testsprite_tests/requirements.txt
requests
python-socketio
websocket-client
pytest
